requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
frozendict==2.3.10
humanize==4.9.0
idna==3.7
isodate==0.6.1
lxml==5.2.1
oauthlib==3.2.2
onelogin==3.1.6
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.7.5
python3-openid==3.2.0
python3-saml==1.16.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.3.0
urllib3==1.26.18
xmlsec==1.3.14
